refactor: 공지 엔티티 패키지 변경
