+++ src/main/java/shop/tryit/domain/notice/NoticeRepository.java
@@ -3,6 +3,7 @@ package shop.tryit.domain.notice;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import shop.tryit.domain.notice.entity.Notice;
 
 public interface NoticeRepository {
 
@@ -13,5 +14,5 @@ public interface NoticeRepository {
     void delete(Notice notice);
 
     Page<Notice> searchNotices(Pageable pageable);
-    
+
 }
