+++ src/main/java/shop/tryit/domain/notice/NoticeService.java
@@ -7,6 +7,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import shop.tryit.domain.notice.entity.Notice;
 
 @Service
 @Transactional(readOnly = true)
