+++ src/main/java/shop/tryit/repository/notice/NoticeJpaRepository.java
@@ -1,7 +1,7 @@
 package shop.tryit.repository.notice;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.tryit.domain.notice.Notice;
+import shop.tryit.domain.notice.entity.Notice;
 
 public interface NoticeJpaRepository extends JpaRepository<Notice, Long> {
