+++ src/test/java/shop/tryit/domain/notice/NoticeServiceTests.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
+import shop.tryit.domain.notice.entity.Notice;
 
 @Transactional
 @SpringBootTest
